fix: remove MeasureGroupHeader.totalItems
Resolves #549
