--- src/test/resources/executeafm/response/measureGroupHeaderFull.json
@@ -1,20 +1,0 @@
-{
-  "measureGroupHeader": {
-    "items": [
-      {
-        "measureHeaderItem": {
-          "name": "Name",
-          "format": "#,##0.00",
-          "localIdentifier": "m1"
-        }
-      }
-    ],
-    "totalItems": [
-      {
-        "totalHeaderItem": {
-          "name": "sum"
-        }
-      }
-    ]
-  }
-}
